requirements.txt
torch
torchvision
numpy
pandas
Pillow
matplotlib
scikit-learn
argparse